+++ Assets/Scripts/Game/FM/FM_Item_Text.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a042148932aa5b54a910451370e6cb55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
